% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateStrandCoverage.R
\name{calculateStrandCoverage}
\alias{calculateStrandCoverage}
\title{Calculate the strand information based on coverage}
\usage{
calculateStrandCoverage(winPositiveAlignments, winNegativeAlignments,
 = 1000, winStep = 100)
}
\arguments{
\item{winPositiveAlignments}{a list that has a `Coverage` field containing 
coverage coming from positive reads}

\item{winNegativeAlignments}{a list that has a `Coverage` field containing 
coverage coming from negative reads}

\item{winWidth}{the length of the sliding window, 1000 by default.}

\item{winStep}{the step length to sliding the window, 100 by default.}
}
\value{
a list of two vectors, containing a positive/negative coverage of 
the input positive/negative windows
}
\description{
Calculate the coverage coming from '+'/'-' reads 
in all sliding wndows
}
