% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateStrandNbReads.R
\name{calculateStrandNbReads}
\alias{calculateStrandNbReads}
\title{Calculate the strand information based the number of reads}
\usage{
calculateStrandNbReads(winPosAlignments, winNegAlignments)
}
\arguments{
\item{winPosAlignments}{a list that has a `Win` field that contains 
information of sliding windows overalapping positive reads}

\item{winNegAlignments}{a a list that has a `Win` field that contains 
information of sliding windows overalapping negative reads}
}
\value{
a list of two vectors, containing a positive/negative number of 
reads of the input positive/negative windows
}
\description{
Calculate the number of reads coming from '+'/'-' strands in 
all sliding wndows
}
