% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromosomeInfoInPartition.R
\name{chromosomeInfoInPartition}
\alias{chromosomeInfoInPartition}
\title{calculate the information within parition}
\usage{
chromosomeInfoInPartition(chromosomeInfo, winStep)
}
\arguments{
\item{chromosomeInfo}{a data frame that contains some information of the alignments}

\item{winStep}{the winStep of sliding windows}
}
\description{
update the information of first/base base/read of each chromosome within the partition
}
