% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computeWin}
\alias{computeWin}
\title{compute strand information of sliding window based on coverage}
\usage{
computeWin(covPosLen, covPosVal, covNegLen, covNegVal, readLength, end, win,
  step, minCov, maxCov, logitThreshold)
}
\arguments{
\item{covPosLen}{the length the Rle object, which is the coverage comes from positive reads}

\item{covPosVal}{the value the Rle object, which is the coverage comes from positive reads}

\item{covNegLen}{the length the Rle object, which is the coverage comes from negative reads}

\item{covNegVal}{the value the Rle object, which is the coverage comes from negative reads}

\item{readLength}{the average length of reads}

\item{end}{the last base on the reference chromosome that the sliding window atteint}

\item{win}{the size of the sliding window}

\item{step}{the step of the sliding window}

\item{minCov}{if a window has the max coverage least than minCov, then it will be rejected}

\item{maxCov}{if a window has the max coverage greater than maxCov, then it will be kept}

\item{logitThreshold}{the logit of the threshold}
}
\value{
A list of two data frames Plus and Minus which respectively contains information of positive windows and negative windows: 'win' is the window number, and 'value' is the normalized estimated value to be tested
}
\description{
non
}
\examples{
computeWin(c(10000,200,30,1,20,50),c(0,2,3,4,1,2),c(10020,300,20,1,10,15),c(0,1,3,5,1,3),100,10200,1000,100,0,0,0.8472979)

}

