% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatenateAlignments.R
\name{concatenateAlignments}
\alias{concatenateAlignments}
\title{concatenate a list of alignments into one}
\usage{
concatenateAlignments(bam, nbOriginalReadsInPart, lengthSeqInPart, nbAlignments,
  flag = FALSE, qname = FALSE)
}

