% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatenateAlignments.R
\name{.concatenateAlignments}
\alias{.concatenateAlignments}
\title{Concatenate a list of Alignments into One}
\usage{
.concatenateAlignments(readInfo, seqInfo)
}
\arguments{
\item{readInfo}{a list returned by \code{scanBam} function, each element 
correspond to a sequence, containing the information of strand, starting 
position, cigar string, and eventually flag, qname}

\item{seqInfo}{a data frame that contains some key information of the 
alignments}
}
\value{
the concatenated alignments of the input list
}
\description{
Concatenate a list of Alignments from multiple sequences into a 
single object
}
\details{
This method take a list of alignments across one or more sequences 
as output by \code{scanBam} 
and concatenates them into a single set of alignments which may include 
multiple sequences
}
\keyword{internal}
