% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMustKeepWinId.R
\name{.getMustKeepWinId}
\alias{.getMustKeepWinId}
\title{Get the Sliding Windows that overlap a GRanges object}
\usage{
.getMustKeepWinId(x, seqInfo, winWidth = 1000L, winStep = 100L)
}
\arguments{
\item{x}{a GRanges object, which defines the coordinates of 
the ranges in the reference genome that all reads mapped to those ranges
must be kept by the filtering method \code{filterDNA}.}

\item{seqInfo}{a data frame that contains some key information of 
the alignments.}

\item{winWidth}{the width of the sliding window, 1000 by default.}

\item{winStep}{the step length to sliding the window, 100 by default.}
}
\value{
A list of two logical vectors (for positive and negative strand) 
defining which windows that overlap the given Granges objects
}
\description{
Get the sliding windows that overlap a GRanges object.
}
\details{
This finds the windows that overlaps the positive/negative strand of a 
GRanges object. The GRanges object, which is \code{mustKeepRanges} in the 
\code{filterDNA} method, defines the coordinates of the ranges in the 
reference genome that all reads mapped to those ranges must be kept by the 
filtering method \code{filterDNA}. 
This method makes use of the method \code{getWinIdOverlapIRanges} by 
pretending each given range as the range of a read, and using readProp=0 
because we want to keep as much reads as possible - so every read that 
overlaps with the given \code{mustKeepRanges}, even by one base.
}
\keyword{internal}
