% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterDNAPairs.R
\name{filterDNAPairs}
\alias{filterDNAPairs}
\title{Filter Paired End Bam File}
\usage{
filterDNAPairs(bamfilein, bamfileout, statfile, chromosomes,
  yieldSize = 1e+08, mustKeepRanges, getWin = FALSE, win = 1000,
  step = 100, threshold = 0.7, pvalueThreshold = 0.05, minReads = 0,
  maxReads = 0, errorRate = 0.01, limit = 0.75, pair = "free",
  coverage = FALSE)
}
\arguments{
\item{bamfilein}{the input bam file to be filterd. Your bamfile should be sorted and have an index file located at the same path as well.}

\item{bamfileout}{the output filtered bam file}

\item{statfile}{the file to write the summary of the results}

\item{chromosomes}{the list of chromosomes to be filtered}

\item{yieldSize}{by default is 1e8, i.e. the bam file is read by block of chromosomes such that the total length of each block is at least 1e8}

\item{mustKeepRanges}{a GRanges object defines the ranges such that every read maps to those ranges must be always kept regardless the strand proportion of the windows containing them.}

\item{getWin}{if TRUE, the function will return a data frame containing the information of all windows. It's FALSE by default.}

\item{win}{the length of the sliding window, 1000 by default.}

\item{step}{the step length to sliding the window, 100 by default.}

\item{threshold}{the threshold upper which we keep the reads. 0.7 by default}

\item{pvalueThreshold}{the threshold for the p-value. 0.05 by default}

\item{minReads}{if a window has least than minReads reads, then it will be rejected regardless the strand proportion. 0 by default}

\item{maxReads}{if a window has more than maxReads reads, then it will be kept regardless the strand proportion. If 0 then it doesn't have effect on selecting window. 0 by default.}

\item{errorRate}{the probability that an RNA read takes the false strand. 0.01 by default}

\item{limit}{a read is considered to be included in a window if and only if at least limit percent of it is in the window. 0.75 by default}

\item{pair}{if "free" than the any pair of reads do not need to be both kept in the filtered file, i.e. first reads and second reads are filtered independently. Otherwise, if pair = "intersection" then a pair of reads is kept if and only if both first and second reads are kept; if pair = "union" then a pair of reads is kept if either the first or the second read is kept.}

\item{coverage}{if TRUE, then the strand information in each window corresponds to the sum of coverage coming from positive/negative reads; and not the number of positive/negative reads as default.}
}
\description{
Filter putative double strand DNA from a strand specific RNA-seq using a window sliding across the genome.
}
\details{
filterDNA reads a bam file containing strand specific RNA reads, and filter putative double strand DNA.
Using a window sliding across the genome, we calculate the positive/negative proportion of reads in that window.
For each window, we use logistic regression to estimate the proportion of reads in the window derived from
stranded RNA (positive or negative).
}
\examples{
bamfilein <- system.file("data","120.10.bam",package = "rnaCleanR")
filterDNAPairs(bamfilein,bamfileout="out.bam",statfile = "out.stat")
}
\seealso{
filterDNA, getWin, getWinPairs, plotHist, plotWin
}

