% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterMultiCount.R
\name{filterMultiCount}
\alias{filterMultiCount}
\title{filter several bamfiles based on the strand specific of all of them. Strand of sliding windows are caculated based on read counts.}
\usage{
filterMultiCount(bamfilein, bamfileout, chromosomes = NULL, win = 1000,
  step = 100, threshold, pvalueThreshold = 0.05, minR = 0, limit = 0.25)
}
\arguments{
\item{bamfilein}{the input bam files to be filterd}

\item{bamfileout}{the output bam files}

\item{chromosomes}{the list of chromosomes to be filtered}

\item{win}{the length of the sliding window}

\item{step}{the step length to slide the window}

\item{threshold}{the threshold upper which we keep the reads}

\item{pvalueThreshold}{the threshold for the p-value}
}
\description{
filter several bamfiles based on the strand specific of all of them. Strand of sliding windows are caculated based on read counts.
}

