% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterMultiCov.R
\name{filterMultiCov}
\alias{filterMultiCov}
\title{Filter Using Merged Files}
\usage{
filterMultiCov(bamfilein, bamfileout, chromosomes = NULL, win = 1000,
  step = 100, pvalueThreshold = 0.05, minR = 0, maxR = 0,
  threshold = 0.7)
}
\arguments{
\item{bamfilein}{the input bam files to be filterd}

\item{bamfileout}{the output bam files}

\item{chromosomes}{the list of chromosomes to be filtered}

\item{win}{the length of the sliding window}

\item{step}{the step length to slide the window}

\item{pvalueThreshold}{the threshold for the p-value}

\item{minR}{if a window has the max coverage least than minR, then it will be rejected}

\item{maxR}{if a window has the max coverage greater than maxR, then it will be kept}

\item{threshold}{the threshold upper which we keep the reads}
}
\value{
the proportion of removed reads in every sample
}
\description{
Filter several bamfiles based on the strand specific of the merged files. 
The strand of each sliding window is caculated based on coverage.
}
\details{
None
}

