% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getIndex}
\alias{getIndex}
\title{Get the indices of positive and negative reads in a given alignment}
\usage{
getIndex(strand)
}
\arguments{
\item{strand}{a string vector contain the "+" or "_"}
}
\value{
A list of two vectors which contains the order of each positive, negative strands within the input strand vector
}
\description{
non
}
\examples{
strand <- rep("+",1000)
negInd <- sample(1:1000,400,replace = FALSE)
strand[negInd] <- "_"
index <- getIndex(strand)


}

