% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlot.R
\name{getPlot}
\alias{getPlot}
\title{Get The Plots Generated From Input Bam File}
\usage{
getPlot(bamfilein, histPlotFile, winPlotFile, chromosomes = NULL, readLength,
  win, step, breaks = 100, xlim, minCov = 0)
}
\arguments{
\item{bamfilein}{the input bam file}

\item{histPlotFile}{the file name of the histogram plot}

\item{winPlotFile}{the file name of the windows plot}

\item{chromosomes}{the list of chromosomes to be filtered}

\item{readLength}{the average length of the reads in the input bam file}

\item{win}{the length of the sliding window}

\item{step}{the step length to slide the window}

\item{breaks}{the breaks of the histogram plot}

\item{xlim}{xlim of the window plot}

\item{minCov}{the min coverage of a window under which the window is ignored}
}
\description{
Plot the histogram of the positive proportion over all sliding windows, and the plot presenting number of reads vs positive proportion over all sliding windows.
}
\examples{
bamfilein <- system.file("data","s1.chr1.bam",package = "rnaCleanR")
getPlot(bamfilein,histPlotFile = "hist.pdf", winPlotFile = "win.pdf",readLength = 100)
}

