% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWin.R
\name{getWin}
\alias{getWin}
\title{Get The Information of Every Window Generated From an Input Single End Bam File}
\usage{
getWin(bamfilein, chromosomes, readLength, win, step, minCov = 0)
}
\arguments{
\item{bamfilein}{the input bam file}

\item{chromosomes}{the list of chromosomes to be considered}

\item{readLength}{the average length of the reads in the input bam file}

\item{win}{the length of the sliding window}

\item{step}{the step length to slide the window}
}
\description{
Return a data frame containing the information of every window generated from the input bam file. The bam file should be single end.
}
\examples{
bamfilein <- system.file("data","s1.chr1.bam",package = "rnaCleanR")
allWin <- getWin(bamfilein,readLength = 50)
}
\seealso{
getWinPairs
}

