% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWinFromPairedBamFile.R
\name{getWinFromPairedBamFile}
\alias{getWinFromPairedBamFile}
\title{get the strand information of all windows from a paired end bam file}
\usage{
getWinFromPairedBamFile(file, chromosomes, mapqFilter = 0,
  partitionSize = 1e+08, winWidth = 1000, winStep = 100, limit = 0.75,
  coverage = FALSE)
}
\arguments{
\item{file}{the input paired-end bam file. Your bamfile should be sorted and have an index file located at the same path as well.}

\item{chromosomes}{the list of chromosomes to be read}

\item{mapqFilter}{every read that has mapping quality below \code{mapqFilter} will be removed before any analysis}

\item{partitionSize}{by default is 1e8, i.e. the bam file is read by block of chromosomes such that the total length of each block is at least 1e8}

\item{winWidth}{the length of the sliding window, 1000 by default.}

\item{winStep}{the step length to sliding the window, 100 by default.}

\item{limit}{a read is considered to be included in a window if and only if at least \code{limit} percent of it is in the window. 0.75 by default}

\item{coverage}{if TRUE, then the strand information in each window corresponds to the sum of coverage coming from positive/negative reads; and not the number of positive/negative reads as default.}
}
\description{
get the number of positive/negative reads of all windows from a paired end bam file
}
\examples{
#' 
bamfilein <- system.file("data","120.10.bam",package = "strandCheckR")
win <- getWinFromPairedBamFile(bamfilein)

}
\seealso{
filterDNA, filterDNAPairs, getWinFromBamFile, plotHist, plotWin
}
