% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWinFromReadInfo.R
\name{getWinFromReadInfo}
\alias{getWinFromReadInfo}
\title{get the strand information of all windows from read information}
\usage{
getWinFromReadInfo(readInfo, winWidth = 1000, winStep = 100,
readProp = 0.5, subset = NULL)
}
\arguments{
\item{readInfo}{a list contains read information returned by 
\code{\link{scanBam}} function when read a bam file}

\item{winWidth}{the width of the sliding window, 1000 by default.}

\item{winStep}{the step length to sliding the window, 100 by default.}

\item{readProp}{A read is considered to be included in a window if at least
\code{readProp} of it is in the window. Specified as a proportion.
0.5 by default.}

\item{subset}{an integer vector specifying the subset of reads to consider}
}
\value{
a DataFrame object containing the number of positive/negative reads 
and coverage of each window sliding
}
\description{
get the number of positive/negative reads of all windows from 
read information obtained from \code{\link{scanBam}} function
}
\examples{
library(Rsamtools)
file <- system.file("extdata","s1.sorted.bam",package = "strandCheckR") 
readInfo <- scanBam(file)
win <- getWinFromReadInfo(readInfo)
}
\seealso{
\code{\link{filterDNA}}, \code{\link{getWinFromBamFile}}
}
