% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWinInChromosome.R
\name{getWinInChromosome}
\alias{getWinInChromosome}
\title{get window data frame with the correct chromosome name and position}
\usage{
getWinInChromosome(Win, chromosomes, chromosomeInfo, winWidth, winStep)
}
\arguments{
\item{Win}{a data frame contains the information of every window in \code{chromosomes}}

\item{chromosomes}{a vector of chromosome names}

\item{chromosomeInfo}{a data frame that contains some information of the alignments}

\item{winWidth}{the length of sliding window}

\item{winStep}{the step length to sliding the window}
}
\description{
get the correct chromsome name and position for each window
}
