% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWinPairs.R
\name{getWinPairs}
\alias{getWinPairs}
\title{Get The Information of Every Window Generated From an Input Paired End Bam File}
\usage{
getWinPairs(bamfilein, chromosomes, readLength, win, step)
}
\arguments{
\item{bamfilein}{the input bam file}

\item{chromosomes}{the list of chromosomes to be considered}

\item{readLength}{the average length of the reads in the input bam file}

\item{win}{the length of the sliding window}

\item{step}{the step length to slide the window}
}
\description{
Return a data frame containing the information of every window generated from the input bam file. The bam file should be paired end.
}
\examples{
bamfilein <- system.file("data","120.10.bam",package = "rnaCleanR")
allWin <- getWinPairs(bamfilein,readLength = 100)
}
\seealso{
getWinPairs
}

