% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histPlot.R
\name{histPlot}
\alias{histPlot}
\title{Plot the histogram of positive proportion of all sliding windows}
\usage{
histPlot(windows, histfile, breaks)
}
\arguments{
\item{windows}{the data frame containing the positive proportion of each window and the coverage group that it belongs to.
By default definition, group can be 1,2...,8 which means its max coverage is respectively in the range "0-10","10-20","20-50","50-100","100-200","200-500","500-1000",">1000"}

\item{histfile}{the file to write the histogram}

\item{breaks}{the number of cells for the histogram}
}
\description{
Plot the histogram of positive proportion of all sliding windows
}
\examples{
windows <- data.frame("propor"=runif(1000, min=0, max=1),"group"=sample(1:8, 1000, replace=TRUE))
histPlot(windows,"hist.pdf",100)
}

