% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keepCount.R
\name{keepCount}
\alias{keepCount}
\title{Calculate the Reads to be Kept}
\usage{
keepCount(bamfilein, chromosomes, allChromosomes, lenSeq, win, step,
  pvalueThreshold, minCov, maxCov, limit, threshold)
}
\arguments{
\item{bamfilein}{the input bam files to be filterd}

\item{chromosomes}{the list of chromosomes to be filtered}

\item{allChromosomes}{the list of all chromosomes}

\item{lenSeq}{the length of every chromosome}

\item{win}{the length of the sliding window}

\item{step}{the step length to slide the window}

\item{pvalueThreshold}{the threshold for the p-value}

\item{minCov}{if a window has the max coverage least than minCov, then it will be rejected}

\item{maxCov}{if a window has the max coverage greater than maxCov, then it will be kept}

\item{limit}{the proportion that a read must be inside a window in order to be counted}

\item{threshold}{the threshold upper which we keep the reads}
}
\description{
None
}

