% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{keepRead}
\alias{keepRead}
\title{Calculate the reads to be kept.}
\usage{
keepRead(posFragments, negFragments, keptPosWin, keptNegWin, win, step,
  errorRate)
}
\arguments{
\item{posFragments}{the data frame contains the information of positive fragments (generated from the function getFragment)}

\item{negFragments}{the data frame contains the information of negative fragments (generated from the function getFragment)}

\item{keptPosWin}{the data frame contains the information of positive kept windows}

\item{keptNegWin}{the data frame contains the information of negative kept windows}

\item{win}{the size of the sliding windows}

\item{step}{the size of the step of sliding windows}

\item{errorRate}{the probability that an RNA read takes the false strand}
}
\value{
A list of two vectors containing the positive and negative alignments to be kept.
}
\description{
Calculate the reads to be kept based of the strand proportion of kept windows. This function is called by the functions filterOne or filterMulti.
If a kept window has more positive alignments than negative alignments, then 
each postive alignment in that window is kept with a probability equal to (positive proportion -  negative proportion)/(positive proportion); and each negative alignment of that window is kept with a probability equal to the given errorRate (0.01 by default). Similarly for windows that have negative alignments more than positive alingments.
Since each alignment can be belonged to several windows, then the final probability for keeping of each alignment is the maximum one from all windows containing it.
}
\examples{
posFragments <- data.frame("group"=sample(1:800,1000,replace=TRUE),"start"=sample(1:1000000,1000,replace=TRUE)) 
posFragments <- posFragments[order(posFragments$start),] \%>\% dplyr::mutate(end=start+sample(50:100,1000,replace=TRUE))
negFragments <- data.frame("group"=sample(1:800,1000,replace=TRUE),"start"=sample(1:1000000,1000,replace=TRUE)) 
negFragments <- negFragments[order(negFragments$start),] \%>\% dplyr::mutate(end=start+sample(50:100,1000,replace=TRUE))
keptPosWin <- data.frame("win"=sample(1:10000,5000,replace=FALSE),"propor"=runif(5000, min=0.7, max=1))
keptPosWin <- keptPosWin[order(keptPosWin$win),]
keptNegWin <- data.frame("win"=sample(1:10000,5000,replace=FALSE),"propor"=runif(5000, min=0, max=0.3))
keptNegWin <- keptNegWin[order(keptNegWin$win),]
win <- 1000
step <- 100
errorRate <- 0.01
reads <- rnaCleanR::keepRead(posFragments,negFragments,keptPosWin,keptNegWin,win,step,errorRate)

}
\seealso{
getFragment, filterOne, filterMulti
}

