% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keptAlignment.R
\name{keptAlignment}
\alias{keptAlignment}
\title{get the kept probability of each read fragment}
\usage{
keptAlignment(fragments, keptProbaW, errorRate)
}
\arguments{
\item{fragments}{an IRange object defind the starting, ending position of each fragment}

\item{keptProbaW}{an Rle object define the kept probability of every sliding window}

\item{errorRate}{the rate that a read has wrong strand}
}
\description{
calculate the keeping probability of each read fragment based on the keeping probability of the windows containing it
}
