% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keptProbaWin.R
\name{keptProbaWin}
\alias{keptProbaWin}
\title{get the probability of keeping each window}
\usage{
keptProbaWin(winPositiveAlignments, winNegativeAlignments, logitThreshold,
  errorRate, mustKeepWin, min, max, getWin, coverage = FALSE)
}

