% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keptReadFragment.R
\name{keptReadFragment}
\alias{keptReadFragment}
\title{calculate the read fragments to be kepts}
\usage{
keptReadFragment(fragments, keptProbaW, errorRate)
}
\arguments{
\item{fragments}{an IRange object defind the starting, ending position of each fragment}

\item{keptProbaW}{an Rle object define the kept probability of each sliding window}

\item{errorRate}{the rate that a read has wrong strand}
}
\description{
calculate the keeping probability of each read fragment based on the keeping probability of the windows containing it. Then get the list of read fragments to be kept.
}
