% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitionSequences.R
\name{partitionSequences}
\alias{partitionSequences}
\alias{partitionSeqinfo}
\title{Partition the sequences}
\usage{
partitionSequences(sequences, lengthSeq, partitionSize = 1e+08)

partitionSeqinfo(sq, partitionSize = 1e+08)
}
\arguments{
\item{sequences}{a vector of sequences to be partitioned}

\item{lengthSeq}{the length of each sequence}

\item{partitionSize}{the minimum sum of length of the sequences in 
each partition}

\item{sq}{a seqinfo object to be partitioned}
}
\value{
a list whose each element is a character vector containing the 
sequence names of a part
}
\description{
Partition the set of sequences for easier processing
}
\details{
Partition the set of sequences such that each part contains a 
subset of sequences whose sum of length is >= \code{partitionSize}.
This is to enable faster processing by \code{getWinFromBamFile} and 
\code{filterDNA}
}
