% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequenceInfoInPartition.R
\name{sequenceInfoInPartition}
\alias{sequenceInfoInPartition}
\title{calculate the information within parition}
\usage{
sequenceInfoInPartition(sequenceInfo, winStep)
}
\arguments{
\item{sequenceInfo}{a data frame that contains some information of the alignments}

\item{winStep}{the winStep of sliding windows}
}
\description{
update the information of first/base base/read of each sequence within the partition
}
