% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statInfoInPartition.R
\name{statInfoInPartition}
\alias{statInfoInPartition}
\title{calculate the information within parition}
\usage{
statInfoInPartition(statInfo, winStep)
}
\arguments{
\item{statInfo}{a data frame that contains some information of the alignments}

\item{winStep}{the winStep of sliding windows}
}
