% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strandCheckR-package.R
\docType{package}
\name{strandCheckR-package}
\alias{strandCheckR}
\alias{strandCheckR-package}
\title{strandCheckR: Check strandedness of alignments in one or more bam files}
\description{
- This package contains functions for checking the strandedness of alignments
such as [getStrandFromBamFile()]
- The results can easily be visualised using [plotHist()]
- Any suspected genomic DNA can also be removed using [filterDNA()], with
thresholds for filtering able to be determined manually using [plotWin()]
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/UofABioinformaticsHub/strandCheckR}
  \item Report bugs at \url{https://github.com/UofABioinformaticsHub/strandCheckR/issues}
}

}
\author{
Thu-Hien To
}
\keyword{internal}
