% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeBam.R
\name{writeBam}
\alias{writeBam}
\title{Write Bam File}
\usage{
writeBam(keep, bamfilein, bamfileout, chromosomes, allChromosomes, lenSeq)
}
\arguments{
\item{keep}{the indices of the reads in the input bamfile to be kept}

\item{bamfilein}{the input bam file}

\item{bamfileout}{the output bam file}

\item{chromosomes}{the list of chromosomes to write}

\item{allChromosomes}{the list of all chromosomes}

\item{lenSeq}{the length of every chromosome}
}
\description{
write the reads into the output bamfile
}
\details{
None
}

