% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatenateAlignments.R
\name{concatenateAlignments}
\alias{concatenateAlignments}
\title{concatenate a list of alignments into one}
\usage{
concatenateAlignments(bam, statinfo, flag = FALSE, qname = FALSE)
}
\arguments{
\item{bam}{a list returned by \code{scanBam} function, each element correspond to a chromosome, containing the information of strand, starting position, cigar string, and eventually flag, qname}

\item{qname}{}
}
\description{
concatenate a list of alignments into one
}
