% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatenateAlignments.R
\name{concatenateAlignments}
\alias{concatenateAlignments}
\title{Concatenate a list of Alignments into One}
\usage{
concatenateAlignments(bam, statInfo)
}
\arguments{
\item{bam}{a list returned by \code{scanBam} function, each element correspond to a chromosome, containing the information of strand, starting position, cigar string, and eventually flag, qname}

\item{statInfo}{a data frame that contains some key information of the alignments}
}
\value{
the concatenated alignments of the input list
}
\description{
Concatenate a list of Alignments from multiple chromosomes into a single object
}
\details{
This method take a list of alignments across one or more chromosomes as output by \code{scanBam} 
and concatenates them into a single set of alignments which may include multiple chromosomes
}
