% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWinInSequence.R
\name{.getWinInSequence}
\alias{.getWinInSequence}
\title{Get window data frame with the correct sequence name and position}
\usage{
.getWinInSequence(Win, seqInfo, winWidth = 1000L, winStep = 100L)
}
\arguments{
\item{Win}{a data frame contains the strand information of every window}

\item{seqInfo}{a data frame that contains some key information of the 
sequences}

\item{winWidth}{the length of sliding window}

\item{winStep}{the step length to sliding the window}
}
\description{
Get the correct sequence name and position for each window
}
\keyword{internal}
