% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequenceInfoInPartition.R
\name{.sequenceInfoInPartition}
\alias{.sequenceInfoInPartition}
\title{Calculate the first/last base/read of each sequence within each part
of the partition.}
\usage{
.sequenceInfoInPartition(seqInfo, winWidth, winStep)
}
\arguments{
\item{seqInfo}{a data frame that contains some key information of the 
sequences}

\item{winWidth}{the length of sliding window}

\item{winStep}{the step length to sliding the window}
}
\description{
Calculate the first/last base/read of each sequence within each part
of the partition.
}
\keyword{internal}
