% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterDNA.R
\name{filterDNA}
\alias{filterDNA}
\title{Filter Single End Bam File}
\usage{
filterDNA(bamfilein, bamfileout, statfile, chromosomes, yieldSize = 1e+08,
  mustKeepRanges, getWin = FALSE, win = 1000, step = 100,
  threshold = 0.7, pvalueThreshold = 0.05, min = 0, max = 0,
  errorRate = 0.01, limit = 0.75, coverage = FALSE)
}
\arguments{
\item{bamfilein}{the input single end bam file to be filterd. Your bamfile should be sorted and have an index file located at the same path as well.}

\item{bamfileout}{the output filtered bam file}

\item{statfile}{the file to write the summary of the results}

\item{chromosomes}{the list of chromosomes to be filtered}

\item{yieldSize}{by default is 1e8, i.e. the bam file is read by block of chromosomes such that the total length of each block is at least 1e8}

\item{mustKeepRanges}{a GRanges object defines the ranges such that every read maps to those ranges must be always kept regardless the strand proportion of the windows containing them.}

\item{getWin}{if TRUE, the function will return a data frame containing the information of all windows. It's FALSE by default.}

\item{win}{the length of the sliding window, 1000 by default.}

\item{step}{the step length to sliding the window, 100 by default.}

\item{threshold}{the threshold upper which we keep the windows. 0.7 by default}

\item{pvalueThreshold}{the threshold for the p-value. 0.05 by default}

\item{min}{if a window has least than \code{min} reads, then it will be rejected regardless the strand proportion. 0 by default}

\item{max}{if a window has more than \code{max} reads, then it will be kept regardless the strand proportion. If 0 then it doesn't have effect on selecting window. 0 by default.}

\item{errorRate}{the probability that an RNA read takes the false strand. 0.01 by default}

\item{limit}{a read is considered to be included in a window if and only if at least \code{limit} percent of it is in the window. 0.75 by default}

\item{coverage}{if TRUE, then the strand information in each window corresponds to the sum of coverage coming from positive/negative reads; and not the number of positive/negative reads as default.}
}
\description{
Filter putative double strand DNA from a strand specific single-end RNA-seq using a window sliding across the genome.
}
\details{
filterDNA reads a single end bam file containing strand specific RNA reads, and filter putative double strand DNA.
Using a window sliding across the genome, we calculate the positive/negative proportion of reads in that window.
For each window, we use logistic regression to estimate the proportion of reads in the window derived from
stranded RNA (positive or negative).

Let \eqn{\pi} be the proportion of reads in the window derived from stranded RNA (positive or negative)

Null hypothesis: \eqn{\pi < {\pi}_{0}} where \eqn{{\pi}_{0}} is the given threshold.

Only windows with p-value <= 0.05 are kept. For a positive window that is kept, let P be its number of positive reads, and let M
be its number of negative reads. Since these M negative reads should come from double-strand DNA, then there should be also M postive reads among the
P positive reads come from double-strand DNA. In other words, there are only (P-M) positive reads come from RNA. Hence, each positive read in this window is kept
with the probability equalling (P-M)/P. Each negative read is kept with the probability equalling the given \code{errorRate} which is the rate that
an RNA read of your sample has wrong strand.

Since each alignment can be belonged to several windows, then the probability of keeping an alignment is the maximum probability defined by
all windows that contain it.
}
\examples{
bamfilein <- system.file("data","s1.chr1.bam",package = "rnaCleanR")
filterDNA(bamfilein,bamfileout="out.bam",statfile = "out.stat")
}
\seealso{
filterDNAPairs, getWinFromBamFile, getWinFromPairedBamFile, plotHist, plotWin
}
