% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterMulti.R
\name{filterMulti}
\alias{filterMulti}
\title{Filter Several Bam Files}
\usage{
filterMulti(bamfilein, bamfileout, statfile, chromosomes, mustKeepRanges,
  readLength, win, step, pvalueThreshold = 0.05, minCov = 0, maxCov = 0,
  threshold = 0.7, errorRate = 0.01)
}
\arguments{
\item{bamfilein}{the input bam files to be filterd}

\item{bamfileout}{the output bam files}

\item{chromosomes}{the list of chromosomes to be filtered}

\item{mustKeepRanges}{a GRanges object defines the ranges such that every read maps to those ranges must be always kept regardless the strand proportion of the windows containing them.}

\item{win}{the length of the sliding window}

\item{step}{the step length to slide the window}

\item{pvalueThreshold}{the threshold for the p-value}

\item{threshold}{the threshold upper which we keep the reads}

\item{minR}{if a window has the max coverage least than minR, then it will be rejected}

\item{maxR}{if a window has the max coverage greater than maxR, then it will be kept}
}
\description{
Filter several bam files based on the strand specific of the merged files. 
The strand of each sliding window is calculated based on coverage.
}
\details{
filterMulti reads a set of bam files containing strand specific RNA reads, and filter the potential double strand contamination DNA from all these files. 
Similar to the filterOne method, it uses a window sliding across the whole genome, but on the merged alignments from all input bam file.
See filterOne for further information.
}
\examples{
 
bamfileins <- system.file("data",c("s1.chr1.bam","s2.chr1.bam"),package = "rnaCleanR")
filterMulti(bamfileins,bamfileout=c("s1.filter.bam","s2.filter.bam"),statfile = "out.stat", readLength = 100, threshold = 0.7)
}
\seealso{
filterOne, getPlot
}

