% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterMultiCov.R
\name{filterMultiCov}
\alias{filterMultiCov}
\title{filter several bamfiles based on the strand specific of the merged files. Strand of sliding windows are caculated based on coverage.}
\usage{
filterMultiCov(bamfilein, bamfileout, chromosomes, win, step, threshold,
  pvalueThreshold, minR)
}
\arguments{
\item{bamfilein}{the input bam files to be filterd}

\item{bamfileout}{the output bam files}

\item{chromosomes}{the list of chromosomes to be filtered}

\item{win}{the length of the sliding window}

\item{step}{the step length to slide the window}

\item{threshold}{the threshold upper which we keep the reads}

\item{pvalueThreshold}{the threshold for the p-value}
}
\description{
filter several bamfiles based on the strand specific of the merged files. Strand of sliding windows are caculated based on coverage.
}

