% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterOne.R
\name{filterOne}
\alias{filterOne}
\title{Filter One Bam File}
\usage{
filterOne(bamfilein, bamfileout, statfile, chromosomes, mustKeepRanges,
  histPlot = FALSE, winPlot = FALSE, histPlotFile, winPlotFile, readLength,
  win, step, pvalueThreshold = 0.05, minCov = 0, maxCov = 0,
  breaks = 100, threshold = 0.7, errorRate = 0.01)
}
\arguments{
\item{bamfilein}{the input bam file to be filterd. Your bamfile should be sorted and have an index file located at the same path as well.}

\item{bamfileout}{the output filtered bam file}

\item{statfile}{the file to write the summary of the results}

\item{chromosomes}{the list of chromosomes to be filtered}

\item{mustKeepRanges}{a GRanges object defines the ranges such that every read maps to those ranges must be kept.}

\item{histPlot}{if TRUE then a histogram of positive proportion over all window will be generated. It's FALSE by default.}

\item{winPlot}{if TRUE then a plot of sum vs positive proportion over all window will be generated. It's FALSE by default.}

\item{histPlotFile}{the file to write the histogram plot when histPlot is TRUE}

\item{winPlotFile}{the file to write the window plot when winPlot is TRUE}

\item{readLength}{the average length of the reads}

\item{win}{the length of the sliding window}

\item{step}{the step length to sliding the window}

\item{pvalueThreshold}{the threshold for the p-value. 0.05 by default}

\item{minCov}{if a window has the max coverage least than minCov, then it will be rejected regardless the strand proportion. 0 by default}

\item{maxCov}{if a window has the max coverage greater than maxCov, then it will be kept regardless the strand proportion. If 0 then it doesn't have effect on selecting window. 0 by default.}

\item{breaks}{the breaks of histogram when histPlot is TRUE}

\item{threshold}{the threshold upper which we keep the reads. 0.7 by default}

\item{errorRate}{the probability that an RNA read takes the false strand. 0.01 by default}
}
\description{
Filter putative double strand DNA from a strand specific RNA-seq using a window sliding across the genome.
}
\details{
filterOne reads a bam file containing strand specific RNA reads, and filter putative double strand DNA. 
Using a window sliding across the genome, we calculate the positive/negative proportion of reads in that window.
Each alignment will be kept with a probability depending on the positive/negative proportions of all windows containing that alignment.
}
\examples{
 
bamfilein <- system.file("data","s1.chr1.bam",package = "rnaCleanR")
filterOne(bamfilein,bamfileout="out.bam",statfile = "out.stat",histPlot = TRUE,histPlotFile = "hist.pdf", readLength = 100, threshold = 0.7)
}
\seealso{
getPlot, filterMulti
}

