% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterOne.R
\name{filterOne}
\alias{filterOne}
\title{Filter One Bam File}
\usage{
filterOne(bamfilein, bamfileout, statfile, chromosomes, histPlot = FALSE,
  winPlot = FALSE, histPlotFile, winPlotFile, readLength, win, step,
  pvalueThreshold = 0.05, minCov = 0, maxCov = 0, breaks = 100,
  threshold = 0.7, errorRate = 0.01)
}
\arguments{
\item{bamfilein}{the input bam file to be filterd. Your bamfile should be sorted and have an index file located at the same path as well.}

\item{bamfileout}{the output filtered bam file}

\item{statfile}{the file to write the summary of the results}

\item{chromosomes}{the list of chromosomes to be filtered}

\item{histPlot}{if it's TRUE then a histogram plot about the positive proportion over all window will be generated. It's FALSE by default.}

\item{winPlot}{if it's TRUE then a plot of sum vs positive proportion over all window will be generated. It's FALSE by default.}

\item{histPlotFile}{the file to write the histogram plot when histPlot is TRUE}

\item{winPlotFile}{the file to write the window plot when winPlot is TRUE}

\item{win}{the length of the sliding window}

\item{step}{the step length to sliding the window}

\item{pvalueThreshold}{the threshold for the p-value. 0.05 by default}

\item{minCov}{if a window has the max coverage least than minCov, then it will be rejected regardless the strand proportion. 0 by default}

\item{maxCov}{if a window has the max coverage greater than maxCov, then it will be kept regardless the strand proportion. If 0 then it doesn't have effect on selecting window. 0 by default.}

\item{breaks}{the breaks the histogram when histPlot is TRUE}

\item{threshold}{the threshold upper which we keep the reads. 0.7 by default}

\item{errorRate}{the probability that an RNA read takes the false strand}
}
\description{
Filter DNA from a strand specific RNA-seq using a window sliding across the genome. Positive/negative strands of sliding windows are caculated based on the coverage of positive/negative reads that overlap the window.
}
\details{
filterOne reads a bam file containing strand specific RNA reads, and filter the potential double strand contamination DNA from that. 
Using a window sliding across the genome, we estimate the probability to keep or reject it based on the proportion of positive/negative reads in the window.
}
\examples{
 
bamfilein <- system.file("data","s1.chr1.bam",package = "rnaCleanR")
filterOne(bamfilein,bamfileout="out.bam",statfile = "out.stat",histPlot = TRUE,histfile = "hist.pdf", readLength = 100, threshold = 0.7)
}

