% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterOneCov.R
\name{filterOneCov}
\alias{filterOneCov}
\title{filter one bamfile based on the strand specifict of itself. Strand of sliding windows are caculated based on coverage.}
\usage{
filterOneCov(bamfilein, bamfileout, chromosomes = NULL, win = 1000,
  step = 100, pvalueThreshold = 0.05, minR = 0, maxR = 0, threshold)
}
\arguments{
\item{bamfilein}{the input bam file to be filterd}

\item{bamfileout}{the output bam file}

\item{chromosomes}{the list of chromosomes to be filtered}

\item{win}{the length of the sliding window}

\item{step}{the step length to slide the window}

\item{pvalueThreshold}{the threshold for the p-value}

\item{minR}{if a window has least than minR reads then it will be all cleaned}

\item{threshold}{the threshold upper which we keep the reads}
}
\description{
filter one bamfile based on the strand specifict of itself. Strand of sliding windows are caculated based on coverage.
}

