% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterOneCov.R
\name{filterOneCov}
\alias{filterOneCov}
\title{Filter One Bam File Based On Coverage}
\usage{
filterOneCov(bamfilein, bamfileout, chromosomes = NULL, win = 1000,
  step = 100, pvalueThreshold = 0.05, minCov = 0, maxCov = 0,
  threshold = 0.7)
}
\arguments{
\item{bamfilein}{the input bam file to be filterd}

\item{bamfileout}{the output bam file}

\item{chromosomes}{the list of chromosomes to be filtered}

\item{win}{the length of the sliding window}

\item{step}{the step length to slide the window}

\item{pvalueThreshold}{the threshold for the p-value}

\item{minCov}{if a window has the max coverage least than minCov, then it will be rejected}

\item{maxCov}{if a window has the max coverage greater than maxCov, then it will be kept}

\item{threshold}{the threshold upper which we keep the reads}
}
\description{
filter one bamfile based on the strand specifict of itself. 
Strand of sliding windows are caculated based on the sum of coverage.
}
\details{
None
}

