% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWin.R
\name{getWin}
\alias{getWin}
\title{Get The Information of Every Window Generated From Input Bam File}
\usage{
getWin(bamfilein, chromosomes, readLength, win, step)
}
\arguments{
\item{bamfilein}{the input bam file}

\item{chromosomes}{the list of chromosomes to be considered}

\item{readLength}{the average length of the reads in the input bam file}

\item{win}{the length of the sliding window}

\item{step}{the step length to slide the window}

\item{histPlotFile}{the file name of the histogram plot}

\item{winPlotFile}{the file name of the windows plot}
}
\examples{
bamfilein <- system.file("data","s1.chr1.bam",package = "rnaCleanR")
allWin <- getWin(bamfilein,readLength = 100)
}

