% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWinFromBamFile.R
\name{getWinFromBamFile}
\alias{getWinFromBamFile}
\title{get the strand information of all windows from bam files}
\usage{
getWinFromBamFile(files, sequences, mapqFilter = 0, partitionSize = 1e+08,
 = 1000, winStep = 100, readProp = 0.5, paired)
}
\arguments{
\item{files}{the input bam files. Your bamfiles should be sorted and have 
their index files located at the same path.}

\item{sequences}{the list of sequences to be read}

\item{mapqFilter}{every read that has mapping quality below 
\code{mapqFilter} will be removed before any analysis}

\item{partitionSize}{by default is 1e8, i.e. the bam file is read by blocks 
of sequences such that the total length of each block is at least 1e8}

\item{winWidth}{the width of the sliding window, 1000 by default.}

\item{winStep}{the step length to sliding the window, 100 by default.}

\item{readProp}{A read is considered to be included in a window if at least
\code{readProp} of it is in the window. Specified as a proportion.
0.5 by default.}

\item{paired}{if TRUE then the input bamfile will be considered as paired 
end reads. If missing, 100 thousands first reads will be inspected to test 
if the input bam file in paired end or single end.}
}
\value{
a DataFrame object containing the number of positive/negative reads 
and coverage of each window sliding across the bam file
}
\description{
get the number of positive/negative reads of all windows from 
bam files
}
\examples{
library(strandCheckR)
file <- system.file("extdata","s1.sorted.bam",package = "strandCheckR")
win <- getWinFromBamFile(file)
win
}
\seealso{
\code{\link{filterDNA}}, \code{\link{plotHist}}, 
\code{\link{plotWin}}
}
