% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWinFromGRanges.R
\name{getWinFromGRanges}
\alias{getWinFromGRanges}
\title{Get the Sliding Windows that overlap a GRanges object}
\usage{
getWinFromGRanges(x, seqInfo, winWidth = 1000L, winStep = 100L)
}
\arguments{
\item{x}{a GRanges object, which defines the coordinates of 
the ranges in the reference genome that all reads mapped to those ranges
must be kept in the filtering method.}

\item{seqInfo}{a data frame that contains some key information of 
the alignments}

\item{winWidth}{the width of the sliding window, 1000 by default.}

\item{winStep}{the step length to sliding the window, 100 by default.}
}
\value{
A list of two logical vectors (for positive and negative strand) 
defining which windows that overlap the given Granges objects
}
\description{
Get the sliding windows that overlap a GRanges object.
}
\details{
This finds the windows that overlaps the positive/negative strand of a 
GRanges object. The GRanges object, which is \code{mustKeepRanges} in the 
\code{filterDNA} method, defines the coordinates of the ranges in the 
reference genome that all reads mapped to those ranges must be kept in the 
filtering method. 
This method makes use of the \code{getWinFromIRanges} method by pretending 
the given range as the range of a read, and giving readProp=0 because we 
want to keep the reads as much as possible - so every read that overlaps 
even by one base with the given \code{mustKeepRanges}.
}
