% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWinFromGranges.R
\name{getWinFromGranges}
\alias{getWinFromGranges}
\title{get the ranges of windows from a GRanges}
\usage{
getWinFromGranges(gr, chromosomes, statinfo, win, step)
}
\arguments{
\item{gr}{a GRanges object}

\item{chromosomes}{a list of chromosome names}

\item{statinfo}{a data frame that contains some information of the alignments}

\item{win}{the length of sliding window}

\item{step}{the step length to sliding the window}
}
\description{
get the lists of positive/negative windows that overlap a GRanges objects
}
