% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWinFromGranges.R
\name{getWinFromGranges}
\alias{getWinFromGranges}
\title{Get the Ranges of Sliding Windows from a GRanges object}
\usage{
getWinFromGranges(x, chromosomes, chromosomeInfo, winWidth = 1000,
  winStep = 100)
}
\arguments{
\item{x}{a GRanges object}

\item{chromosomes}{a list of chromosome names}

\item{chromosomeInfo}{a data frame that contains some information of the alignments}

\item{winWidth}{The width of each window}

\item{winStep}{The step size for sliding the window}
}
\value{
A list of two logical vectors defining whether to keep windows based on positive or negative strands
}
\description{
Get the Ranges of positive/negative windows that overlap a GRanges object
}
