% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWinInChromosome.R
\name{getWinInChromosome}
\alias{getWinInChromosome}
\title{get window data frame with the correct chromosome name and position}
\usage{
getWinInChromosome(Win, chromosomes, statInfo, win, step)
}
\arguments{
\item{Win}{a data frame contains the information of every window in \code{chromosomes}}

\item{statInfo}{a data frame that contains some information of the alignments}

\item{win}{the length of sliding window}

\item{step}{the step length to sliding the window}

\item{chromsomes}{a vector of chromosome names}
}
\description{
get the correct chromsome name and position for each window
}
