% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWinOfAlignments.R
\name{getWinOfAlignments}
\alias{getWinOfAlignments}
\title{get the window ranges of alignments}
\usage{
getWinOfAlignments(bam, str, winSize, winStep, limit, useCoverage = FALSE,
  subset)
}
\arguments{
\item{bam}{a list contains the read information of one chromosome}

\item{str}{the considering strand}

\item{winSize}{the window size}

\item{winStep}{the window winStep}

\item{limit}{a read is considered to be included in a window if and only if at least \code{limit} percent of it is in the window.}

\item{useCoverage}{either base on coverage or number of reads}

\item{subset}{if we consider only a subset of the input reads}
}
\description{
calculate the windows that contain each read fragment
}
