% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWinOfAlignments.R
\name{getWinOfAlignments}
\alias{getWinOfAlignments}
\title{get the window ranges that contain each read fragment}
\usage{
getWinOfAlignments(bam, str, winSize, winStep, limit, subset,
  useCoverage = FALSE)
}
\arguments{
\item{bam}{a list contains the read information of one chromosome}

\item{str}{the considering strand}

\item{winSize}{the window size}

\item{winStep}{the window winStep}

\item{limit}{a read is considered to be included in a window if and only if at least \code{limit} percent of it is in the window.}

\item{subset}{if we consider only a subset of the input reads}

\item{useCoverage}{either base on coverage or number of reads}
}
