% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWinOfAlignments.R
\name{getWinOfAlignments}
\alias{getWinOfAlignments}
\title{get the window ranges of alignments}
\usage{
getWinOfAlignments(readInfo, strand, winWidth, winStep, readProp,
 = FALSE, subset = NULL)
}
\arguments{
\item{readInfo}{a list contains the read information of one sequence}

\item{strand}{the considering strand}

\item{winWidth}{the window size}

\item{winStep}{the window winStep}

\item{readProp}{a read is considered to be included in a window if and only 
if at least \code{readProp} 
percent of it is in the window.}

\item{useCoverage}{either base on coverage or number of reads}

\item{subset}{if we consider only a subset of the input reads}
}
\value{
If \code{useCoverage=FALSE}: an IRanges object which contains the 
range of sliding windows that overlap each read fragment. 
If \code{useCoverage=TRUE}: a list of two objects, the first one is the 
later IRanges object, the second one is an integer-Rle object which contains 
the coverage of the input readInfo
}
\description{
calculate the windows that contain each read fragment
}
