% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersectWithFeature.R
\name{intersectWithFeature}
\alias{intersectWithFeature}
\title{Intersect the windows with an annotation data frame to get feature for each window}
\usage{
intersectWithFeature(windows, annotation, getFeatureInfo = FALSE,
  overlapCol = "OverlapFeature", mcolsAnnot, collapse, ...)
}
\arguments{
\item{windows}{data frame containing the strand information of the sliding windows.
Windows can be obtained using the function \code{getWinFromBamFile}.}

\item{annotation}{a Grange object that you want to intersect with your windows. 
It can have mcols which contains the information or features that could be able to integrate into the input windows}

\item{getFeatureInfo}{whether to get the information of features in the mcols of annotation data. 
If FALSE the return windows will have an additional column OverlapAnnot indicating wheather each window overlaps with any range in the annotion data.
If TRUE the return windows will contain the information of features that overlap each window}

\item{mcolsAnnot}{the column names of the mcols of annotation that you want to get information}

\item{collapse}{character which is used collapse multiple features that overlap with a same window into a string.
If missing then we don't collapse them.}

\item{...}{used to pass parameters to GenomicRanges::findOverlaps}
}
\description{
Intersect the windows with an annotation data frame to get feature for each window
}
\examples{
\dontrun{
#for single end bam file
bamfilein = system.file("extdata","s1.sorted.bam",package = "strandCheckR")
windows <- getWinFromBamFile(file = bamfilein)
windows$Chr <- "chr10"
library(TxDb.Hsapiens.UCSC.hg38.knownGene)
annot <- transcripts(TxDb.Hsapiens.UCSC.hg38.knownGene)
windows <- intersectWithFeature(windows,annot)
h <- summarizeHist(windows,group_by = "OverlapFeature")
plotHist(h,facets = "OverlapFeature")
plotWin(h,facets = "OveralapFeature")
}

}
\seealso{
getWinFromBamFile, summarizeHist, plotHist, plotWin
}
