% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{keepRead}
\alias{keepRead}
\title{Calculate the reads to be kept.}
\usage{
keepRead(posFragments, negFragments, keptPosWin, keptNegWin, win, step,
  errorRate)
}
\arguments{
\item{posFragments}{the data frame contains the information of positive fragments (generated from the function getFragment)}

\item{negFragments}{the data frame contains the information of negative fragments (generated from the function getFragment)}

\item{keptPosWin}{the data frame contains the information of positive kept windows}

\item{keptNegWin}{the data frame contains the information of negative kept windows}

\item{win}{the size of the sliding windows}

\item{step}{the size of the step of sliding windows}

\item{errorRate}{the probability that an RNA read takes the false strand}
}
\value{
A list of two vectors containing the positive and negative reads to be kept
}
\description{
Calculate the reads to be kept based of the strand proportion of kept windows.
}
\examples{
posFragments <- data.frame("group"=sample(1:800,1000,replace=TRUE),"start"=sample(1:1000000,1000,replace=TRUE)) 
posFragments <- posFragments[order(posFragments$start),] \%>\% dplyr::mutate(end=start+sample(50:100,1000,replace=TRUE))
negFragments <- data.frame("group"=sample(1:800,1000,replace=TRUE),"start"=sample(1:1000000,1000,replace=TRUE)) 
negFragments <- negFragments[order(negFragments$start),] \%>\% dplyr::mutate(end=start+sample(50:100,1000,replace=TRUE))
keptPosWin <- data.frame("win"=sample(1:10000,5000,replace=TRUE),"propor"=runif(5000, min=0, max=1))
keptNegWin <- data.frame("win"=sample(1:10000,5000,replace=TRUE),"propor"=runif(5000, min=0, max=1))
win <- 1000
step <- 100
errorRate <- 0.01
keepRead(posFragments,negFragments,keptPosWin,keptNegWin,win,step,errorRate)

}

