% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitionChromosomes.R
\name{partitionChromosomes}
\alias{partitionChromosomes}
\title{partition the set of chromosomes such that each part contains a subset of chromosome whose sum of length is at least 1e8}
\usage{
partitionChromosomes(chromosomes, lengthSeq, yieldSize)
}

