% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHist.R
\name{plotHist}
\alias{plotHist}
\title{Plot the histogram of positive proportions of the input windows data frame}
\usage{
plotHist(windows, group = c(10, 100, 1000), save = FALSE,
  file = "hist.pdf", facet_wrap_chromosomes = FALSE)
}
\arguments{
\item{windows}{data frame containing the number of positive/negative reads for each window
Windows can be get by the function \code{getWinFromBamFile} (for single end bam file) or \code{getWinFromPairedBamFile} (for paired end bam file).}

\item{group}{an integer vector that specifies how you want to partition the windows based on the number of reads. By default \code{group} = c(10,100,1000), which means that your windows will be parition into 4 groups, those have number of reads < 10, from 10 to 100, from 100 to 1000, and > 1000}

\item{save}{if TRUE, then the plot will be save into the file given by \code{file} parameter}

\item{file}{the file name to save to plot}

\item{facet_wrap_chromosomes}{if TRUE, then the plots will be splitted by chromosomes. FALSE by default}
}
\description{
Plot the histogram of positive proportions of the input windows data frame
}
\examples{
#for single end bam file
bamfilein = system.file("data/s1.chr1.bam",package = "rnaCleanR")
windows <- getWinFromBamFile(bamfilein)
plotHist(windows)
#for paired end bamfile
bamfilepair = system.file("data/120.10.bam",package = "rnaCleanR")
windowsP <- getWinFromPariedBamFile(bamfilein = "data/120.10.bam")
plotHist(windowsP)
}
\seealso{
getWinFromBamFile, getWinFromPairedBamFile, plotWin
}

