% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHist.R
\name{plotHist}
\alias{plotHist}
\title{Plot the histogram of positive proportions of the input histogram data frame}
\usage{
plotHist(histWin, save = FALSE, file = "hist.pdf", facets = NULL,
  heatmap = FALSE, ...)
}
\arguments{
\item{histWin}{data frame obtained from \code{summarizeHist} function}

\item{save}{if TRUE, then the plot will be save into the file given by \code{file} parameter}

\item{file}{the file name to save to plot}

\item{facets}{colnames of \code{histWin} which will be used to split the plot}

\item{heatmap}{if TRUE, then use heat map to plot the histogram, otherwise use barplot. FALSE by default.}

\item{...}{used to pass parameters to facet_wrap}
}
\description{
Plot the histogram of positive proportions of the input histogram  data frame
}
\examples{
\dontrun{
#for single end bam file
bamfilein = system.file("extdata","s1.sorted.bam",package = "strandCheckR")
windows  <- getWinFromBamFile(file = bamfilein)
histWin <- summarizeHist(windows)
plotHist(histWin)}

}
\seealso{
summarizeHist, getWinFromBamFile, plotWin
}
