% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWin.R
\name{plotWin}
\alias{plotWin}
\title{Plot the number of reads vs positive proportion of the input windows data frame.}
\usage{
plotWin(windows, group = c(10, 100, 1000), threshold = c(0.6, 0.7, 0.8,
  0.9), pvalue = 0.05, save = FALSE, file = "win.pdf",
  facet_wrap_chromosomes = FALSE, useCoverage = FALSE)
}
\arguments{
\item{windows}{data frame containing the number of positive/negative reads for each window.
Windows can be get by the function \code{getWinFromBamFile} (for single end bam file) or \code{getWinFromPairedBamFile} (for paired end bam file).}

\item{group}{an integer vector that specifies how you want to partition the windows based on the number of reads. By default \code{group} = c(10,100,1000), which means that your windows will be parition into 4 groups, those have number of reads < 10, from 10 to 100, from 100 to 1000, and > 1000}

\item{threshold}{a real vector that specifies which threshold lines you want to draw on the plot. The positive window above the threshold line (or negative window below the threshold line) will be kept if we use that threshold to filter the data. By default, the thresholds 0.6, 0.7, 0.8, 0.9 will be plotted.}

\item{pvalue}{0.05 by default, any window has pvalue below it in the test comparing with a threshold will be kept.}

\item{save}{if TRUE, then the plot will be save into the file given by \code{file} parameter}

\item{file}{the file name to save to plot}

\item{facet_wrap_chromosomes}{if TRUE, then the plots will be splitted by chromosomes. FALSE by default}

\item{useCoverage}{if TRUE then plot the coverage strand information, otherwise plot the number of reads strand information. FALSE by default}
}
\description{
Plot the number of reads vs positive proportion of the input windows data frame.
}
\examples{

#for single end bam file
bamfilein = system.file("extdata","s1.chr1.bam",package = "strandCheckR")
windows <- getWinFromBamFile(bamfilein)
plotWin(windows)
#for paired end bamfile
bamfilepair = system.file("extdata","120.10.bam",package = "strandCheckR")
windowsP <- getWinFromPairedBamFile(bamfilepair)
plotWin(windowsP)

}
\seealso{
getWinFromBamFile,  plotHist
}
