% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statInfoInPartition.R
\name{statInfoInPartition}
\alias{statInfoInPartition}
\title{calculate the information within parition}
\usage{
statInfoInPartition(statinfo, step)
}
\arguments{
\item{step}{the step of sliding windows}
}
